heredoc helper for tests
I think it's easier to understand the tests if they are displaying the
exact code we mean to compile. I'm just a little worried about some
oddities with extra characters making the test code not look _exactly_
like the string we are parsing.

Plus the heredoc hack is cool and I've wanted a way to use it :)